+++ HISTORY.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+
+- [cpp] fix: pinocchio v4 compatibility
+
 ## 0.6.7 (April 9 2026)
 
 - [cpp] Modernize googletest CMake usage (#283)